fix size

+++ strategies/convert_adr.py
@@ -15,12 +15,12 @@ def trade(exchange):
 
     diff = adr_estimate - ori_estimate
     if diff > 0:
-        size = (diff / 0.1) * 2
+        size = int((diff / 0.1) * 2)
         trades.append(("BUY", "VALBZ", int(ori_estimate + 1), size))
         trades.append(("CONVERT", "BUY", "VALE", size))
         trades.append(("SELL", "VALE", int(adr_estimate - 1), size))
     elif diff < 0:
-        size = (-diff / 0.1) * 2
+        size = int((-diff / 0.1) * 2)
         trades.append(("SELL", "VALBZ", int(ori_estimate + 1), size))
         trades.append(("CONVERT", "SELL", "VALE", size))
         trades.append(("BUY", "VALE", int(adr_estimate - 1), size))
